Moved utils functions into transform_utils
