--- pcdet/utils/depth_utils.py
@@ -1,40 +1,0 @@
-import torch
-import math
-
-
-def bin_depths(depth_map, mode, depth_min, depth_max, num_bins, target=False):
-    """
-    Converts depth map into bin indices
-    Args:
-        depth_map: (H, W), Depth Map
-        mode: string, Discretiziation mode (See https://arxiv.org/pdf/2005.13423.pdf for more details)
-            UD: Uniform discretiziation
-            LID: Linear increasing discretiziation
-            SID: Spacing increasing discretiziation
-        depth_min: float, Minimum depth value
-        depth_max: float, Maximum depth value
-        num_bins: int, Number of depth bins
-        target: bool, Whether the depth bins indices will be used for a target tensor in loss comparison
-    Returns:
-        indices: (H, W), Depth bin indices
-    """
-    if mode == "UD":
-        bin_size = (depth_max - depth_min) / num_bins
-        indices = ((depth_map - depth_min) / bin_size)
-    elif mode == "LID":
-        bin_size = 2 * (depth_max - depth_min) / (num_bins * (1 + num_bins))
-        indices = -0.5 + 0.5 * torch.sqrt(1 + 8 * (depth_map - depth_min) / bin_size)
-    elif mode == "SID":
-        indices = num_bins * (torch.log(1 + depth_map) - math.log(1 + depth_min)) / \
-            (math.log(1 + depth_max) - math.log(1 + depth_min))
-    else:
-        raise NotImplementedError
-
-    if target:
-        # Remove indicies outside of bounds
-        mask = (indices < 0) | (indices > num_bins) | (~torch.isfinite(indices))
-        indices[mask] = num_bins
-
-        # Convert to integer
-        indices = indices.type(torch.int64)
-    return indices
