--- pcdet/utils/grid_utils.py
@@ -1,19 +1,0 @@
-import torch
-
-
-def normalize_coords(coords, shape):
-    """
-    Normalize coordinates of a grid between [-1, 1]
-    Args:
-        coords: (..., 3), Coordinates in grid
-        shape: (3), Grid shape
-    Returns:
-        norm_coords: (.., 3), Normalized coordinates in grid
-    """
-    min_n = -1
-    max_n = 1
-    shape = torch.flip(shape, dims=[0])  # Reverse ordering of shape
-
-    # Subtract 1 since pixel indexing from [0, shape - 1]
-    norm_coords = coords / (shape - 1) * (max_n - min_n) + min_n
-    return norm_coords
